Fix crash on name collisions with typing containers
Due to dynamic method dispatch in the TypeRewriter, any type that matched the
names List, Set, Tuple, etc. would crash for missing `__args__`.

We fix this by checking if the module matches, and then using `getattr()`
anyway, in case there was a custom `typing` module used.

test_typing.py is a bit ugly now due to the name conflict.