Add config to disable applying of TypedDict stubs.
Augment the given type rewriter with one that converts TypedDicts to Dicts. This means that, when TypedDicts are enabled, the traces will still be stored as TypedDicts but the types shrunk from the traces will be Dicts for `apply` and TypedDicts for other modes like `stub` or `diff`.

Note that TypedDict class stubs won't be generated for `apply` since that happens after the traced types are shrunk.

Make sure to convert nested TypedDicts correctly by using the above rewriter wherever we need to convert a TypedDicts to a Dict.