Shrink non-total TypedDicts.
Shrink TypedDicts by looking for fields that are present in all the TypedDicts and fields that are present only in some.

When all fields are optional, generate a non-total TypedDict class stub.
When there are both required and optional fields, generate a total base class and a non-total subclass.

Design choices:

+ If a key has different value types in the traced types, then we shrink those types to get its value type. This should not lead to large types because we shrink large union types to Any.

+ If the resulting TypedDict is larger than max_typed_dict_size, then fall back to `Dict[str, ...]`. However, preserve any nested anonymous TypedDicts.

+ Represent required and optional fields as nested TypedDicts.

+ Add `types_equal` to allow equality for `List[TypedDicts(...)]` on 3.6.