[python 3.11] Fix RenderAnnotation's make_builtin_tuple
This has a special case to handle '()', but the test (whether the
'elements' iterable is truthy) no longer works as it always passes.

Instead, always do the join, then test if it's truthy or not, and
return the failsafe '()' if not.

This addresses one test failure in #272:
```
_______________________ TestRenderAnnotation.test_render_annotation[Tuple-Tuple[()]] _______________________
tests/test_stubs.py:212: in test_render_annotation
    assert render_annotation(annotation) == expected
E   AssertionError: assert 'Tuple[]' == 'Tuple[()]'
E     - Tuple[()]
E     ?       --
E     + Tuple[]
```

Appears to be due to
https://github.com/python/cpython/issues/91137

Signed-off-by: Michel Alexandre Salim <michel@michel-slm.name>